backend/services/ingestion/requirements.txt
fastapi
uvicorn[standard]
httpx
pypdf
python-dotenv
python-multipart
pydantic
